feat: implemented movie deletion and creation authorization checks

+++ src/controllers/movies-controller/mutations/delete-movie.ts
@@ -1,11 +1,12 @@
 import { RequestHandler } from 'express';
-import ErrorService, { ServerSetupError } from '../../../services/error-service';
+import ErrorService, { ServerSetupError, ForbiddenError } from '../../../services/error-service';
 import MoviesModel from '../model';
-import { MovieModel } from '../types';
+import { MovieViewModel } from '../types';
+import UserModel from '../auth/model/index';
 
 export const deleteMovie: RequestHandler<
 { id: string | undefined },
-MovieModel | ResponseError,
+MovieViewModel | ResponseError,
 {},
 {}
 > = async (req, res) => {
@@ -13,7 +14,12 @@ MovieModel | ResponseError,
 
   try {
     if (id === undefined) throw new ServerSetupError();
+    if (req.authData === undefined) throw new ServerSetupError();
+    const user = await UserModel.getUserByEmail(req.authData.email);
     const movie = await MoviesModel.getMovie(id);
+
+    if (user.role !== 'ADMIN' && user.id !== movie.owner.id) throw new ForbiddenError();
+
     await MoviesModel.deleteMovie(id);
 
     res.status(200).json(movie);
